First commit. Many projects from an introductory community college class in java plus some of my own projects.
There are three android projects here as well. Stopwatch, First Android, and XMLtest.

The grandest one here is Student Reporter. This was my first attempt at what I would consider
a real program. With a GUI and everything. It was submitted as an final project
for the introductory community college class. This was done above and beyond the
inital project specifications. Which only called for a simple CLI, writing text to a file
and and printing output.